+++ example/bun.config.dart
@@ -1,9 +1,5 @@
 import 'package:spry/config.dart';
 
 void main() {
-  defineSpryConfig(
-    host: '127.0.0.1',
-    port: 4100,
-    target: BuildTarget.bun,
-  );
+  defineSpryConfig(host: '127.0.0.1', port: 4100, target: BuildTarget.bun);
 }