+++ example/routes/_middleware.dart
@@ -1,6 +1,9 @@
 import 'package:spry/spry.dart';
 
 Future<Response> middleware(Event event, Next next) async {
-  event.locals.set(#requestId, DateTime.now().microsecondsSinceEpoch.toString());
+  event.locals.set(
+    #requestId,
+    DateTime.now().microsecondsSinceEpoch.toString(),
+  );
   return next();
 }