+++ sites/spry.medz.dev/snippets/quickstart/routes/about.get.dart
@@ -1,8 +1,5 @@
 import 'package:spry/spry.dart';
 
 Response handler(Event event) {
-  return Response.json({
-    'page': 'about',
-    'method': event.method,
-  });
+  return Response.json({'page': 'about', 'method': event.method});
 }