+++ test/fixtures/generator/output_dir/src/routes/index.dart
@@ -1,3 +1,3 @@
 import 'package:spry/spry.dart';
 
-Response handler(Event event) => Response.text('ok');
+Response handler(Event event) => Response('ok');